Resolved 39 conflicts across the mobskill/effect rework, FATE zone hooks,
Treasure Hunter job-agnostic change, augmented mimic loot, and the
instanced-fight-system engine files. Notable resolutions:

- Reverted an incomplete local WIP mobskill/effect rework (13 mobskill
  scripts + a quest reward block) back to upstream's complete
  implementation; the local version had unfinished TODOs and one file
  (blade_retsu.lua) was accidentally deleted despite still being an
  active mob skill.
- Merged FATE zone hooks with upstream's new expeditionaryForce/Siren's
  Tear zone init calls across 12 zone files (both coexist).
- Kept the job-agnostic Treasure Hunter change (battleutils.cpp,
  enmity_container.cpp) while adopting upstream's Mod/Job namespacing.
- Kept the ALTER_EGO_SKILL_MULTIPLIER trust stat scaling while adopting
  upstream's xi::Mod namespacing (trustutils.cpp).
- Adopted upstream's rewritten AI state-tick loop (ai_container.cpp)
  wholesale; the local side referenced an undeclared `top` variable and
  did not compile, and upstream's finishCurrentState()/resumeNextState()
  redesign already subsumes the edge case the local patch targeted.
- Merged the instanced-fight-system engine files (lua_instance.cpp,
  luautils.h, instanceutils.cpp, zone_instance.h) to keep both the
  OnInstanceCapacityReached callback and CountInstancesOf addition
  alongside upstream's FnRef/xi::ZoneId refactors.
- Kept the dynamic entity-name packet sizing fix while adding upstream's
  door-entity (door_id) packet case (entity_update.cpp).
- Kept the awardItem()-based augmented mimic loot path in treasure_pool.cpp
  over adopting upstream's new ItemClaimTransaction API for that call site;
  note the augmented branch's itemutils::GetItem() call was already
  undefined/non-compiling before this merge (pre-existing bug, not
  introduced here).

Known follow-up: not build-tested locally (no MSVC/vcpkg toolchain
available in this shell). Build before pushing to origin/base.

+++ src/map/packets/char_sync.cpp
@@ -21,7 +21,7 @@
 
 #include "char_sync.h"
 
-#include "entities/charentity.h"
+#include "entities/char_entity.h"
 #include "status_effect_container.h"
 #include "utils/mountutils.h"
 
@@ -57,6 +57,6 @@ CCharSyncPacket::CCharSyncPacket(CCharEntity* PChar)
         ref<uint32>(0x1C)                           = CustomProperties[1]; // Noble Chocobo
     }
 
-    ref<uint8>(0x25) = PChar->jobs.job[PChar->GetMJob()];
+    ref<uint8>(0x25) = PChar->jobs.job[static_cast<uint8>(PChar->GetMJob())];
     ref<uint8>(0x27) = PChar->profile.mhflag & 0x20 ? 1 : 0; // MogExpansionFlag - Is 2nd floor unlocked.
 }
